+++ src/asldro/filters/asl_quantification_filter.py
@@ -62,16 +62,10 @@ class AslQuantificationFilter(BaseFilter):
 
     **Quantification Model**
 
-    The following equations are used to calculate the perfusion rate, depending on the
-    input ``model``:
+    The following equations are used to calculate the perfusion rate, depending
+    on the input ``model``:
 
-    :'whitepaper': equations used are those described in:
-
-        Alsop et al. Recommended implementation of arterial spin-labeled
-        perfusion MRI for clinical applications: A consensus of the ISMRM
-        perfusion study group and the European consortium for ASL in
-        dementia. Magnetic Resonance in Medicine 2014;73:102–116
-        https://doi.org/10.1002/mrm.25197.
+    :'whitepaper': simplified single subtraction equations :cite:p:`Alsop2014`.
 
       * for pCASL/CASL see :class:`AslQuantificationFilter.asl_quant_wp_casl`
       * for PASL see :class:`AslQuantificationFilter.asl_quant_wp_pasl`.
@@ -275,14 +269,7 @@ class AslQuantificationFilter(BaseFilter):
     ) -> np.ndarray:
         r"""
         Performs ASL quantification using the White Paper equation for
-        (pseudo)continuous ASL, as given in
-
-            Alsop et al. Recommended implementation of arterial spin-labeled
-            perfusion MRI for clinical applications: A consensus of the ISMRM
-            perfusion study group and the European consortium for ASL in
-            dementia. Magnetic Resonance in Medicine 2014;73:102–116
-            https://doi.org/10.1002/mrm.25197.
-
+        (pseudo)continuous ASL :cite:p:`Alsop2014`.
 
         .. math::
             &f = \frac{6000 \cdot\ \lambda \cdot (\text{SI}_{\text{control}} - \text{SI}_{\text{label}}) \cdot
@@ -350,14 +337,7 @@ class AslQuantificationFilter(BaseFilter):
     ) -> np.ndarray:
         r"""
         Performs ASL quantification using the White Paper equation for
-        pulsed ASL, as given in
-
-            Alsop et al. Recommended implementation of arterial spin-labeled
-            perfusion MRI for clinical applications: A consensus of the ISMRM
-            perfusion study group and the European consortium for ASL in
-            dementia. Magnetic Resonance in Medicine 2014;73:102–116
-            https://doi.org/10.1002/mrm.25197.
-
+        pulsed ASL :cite:p:`Alsop2014`.
 
         .. math::
             &f = \frac{6000 \cdot\ \lambda \cdot (\text{SI}_{\text{control}}